Move the verification layer
