+++ x/proof/keeper/msg_server_create_claim.go
@@ -37,53 +37,44 @@ func (k msgServer) CreateClaim(
 	}()
 
 	logger := k.Logger().With("method", "CreateClaim")
+	sdkCtx := cosmostypes.UnwrapSDKContext(ctx)
 	logger.Info("creating claim")
 
+	// Basic validation of the CreateClaim message.
 	if err = msg.ValidateBasic(); err != nil {
 		return nil, err
 	}
+	logger.Info("validated the createClaim message")
 
 	// Compare msg session header w/ on-chain session header.
 	var session *sessiontypes.Session
-	session, err = k.queryAndValidateSessionHeader(ctx, msg)
+	session, err = k.queryAndValidateSessionHeader(ctx, msg.GetSessionHeader(), msg.GetSupplierAddress())
 	if err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	// Use the session header from the on-chain hydrated session.
-	sessionHeader := session.GetHeader()
-
-	// Set the session header to the on-chain hydrated session header.
-	msg.SessionHeader = sessionHeader
-
-	// Validate claim message commit height is within the respective session's
-	// claim creation window using the on-chain session header.
-	if err = k.validateClaimWindow(ctx, msg); err != nil {
-		return nil, status.Error(codes.FailedPrecondition, err.Error())
-	}
-
-	logger = logger.
-		With(
-			"session_id", session.GetSessionId(),
-			"session_end_height", sessionHeader.GetSessionEndBlockHeight(),
-			"supplier", msg.GetSupplierAddress(),
-		)
-
-	logger.Info("validated claim")
-
-	// Assign and upsert claim after all validation.
+	// Construct and insert claim
 	claim = types.Claim{
 		SupplierAddress: msg.GetSupplierAddress(),
-		SessionHeader:   sessionHeader,
+		SessionHeader:   session.GetHeader(),
 		RootHash:        msg.GetRootHash(),
 	}
 
-	_, isExistingClaim = k.Keeper.GetClaim(ctx, claim.GetSessionHeader().GetSessionId(), claim.GetSupplierAddress())
+	// Helpers for logging the same metadata throughout this function calls
+	logger = logger.
+		With(
+			"session_id", session.GetSessionId(),
+			"session_end_height", claim.SessionHeader.GetSessionEndBlockHeight(),
+			"supplier", msg.GetSupplierAddress(),
+		)
 
-	k.Keeper.UpsertClaim(ctx, claim)
+	// Validate claim message commit height is within the respective session's
+	// claim creation window using the on-chain session header.
+	if err = k.validateClaimWindow(ctx, claim.SessionHeader, claim.SupplierAddress); err != nil {
+		return nil, status.Error(codes.FailedPrecondition, err.Error())
+	}
 
-	logger.Info("created new claim")
-
+	// Get metadata for the event we want to emit
 	numRelays, err = claim.GetNumRelays()
 	if err != nil {
 		return nil, status.Error(codes.Internal, types.ErrProofInvalidClaimRootHash.Wrap(err.Error()).Error())
@@ -92,6 +83,11 @@ func (k msgServer) CreateClaim(
 	if err != nil {
 		return nil, status.Error(codes.Internal, types.ErrProofInvalidClaimRootHash.Wrap(err.Error()).Error())
 	}
+	_, isExistingClaim = k.Keeper.GetClaim(ctx, claim.GetSessionHeader().GetSessionId(), claim.GetSupplierAddress())
+
+	// Upsert the claim
+	k.Keeper.UpsertClaim(ctx, claim)
+	logger.Info("successfully upserted the claim")
 
 	// Emit the appropriate event based on whether the claim was created or updated.
 	var claimUpsertEvent proto.Message
@@ -113,8 +109,6 @@ func (k msgServer) CreateClaim(
 			},
 		)
 	}
-
-	sdkCtx := cosmostypes.UnwrapSDKContext(ctx)
 	if err = sdkCtx.EventManager().EmitTypedEvent(claimUpsertEvent); err != nil {
 		return nil, status.Error(
 			codes.Internal,
@@ -126,7 +120,6 @@ func (k msgServer) CreateClaim(
 		)
 	}
 
-	// TODO_BETA: return the claim in the response.
 	return &types.MsgCreateClaimResponse{
 		Claim: &claim,
 	}, nil
