+++ x/proof/keeper/msg_server_create_claim_test.go
@@ -5,11 +5,10 @@ import (
 
 	abci "github.com/cometbft/cometbft/abci/types"
 	cosmostypes "github.com/cosmos/cosmos-sdk/types"
+	"github.com/pokt-network/smt"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-
-	"github.com/pokt-network/smt"
 
 	keepertest "github.com/pokt-network/poktroll/testutil/keeper"
 	testproof "github.com/pokt-network/poktroll/testutil/proof"
@@ -140,7 +139,7 @@ func TestMsgServer_CreateClaim_Success(t *testing.T) {
 			events := sdkCtx.EventManager().Events()
 			require.Equal(t, 1, len(events))
 
-			require.Equal(t, events[0].Type, "poktroll.proof.EventClaimCreated")
+			require.Equal(t, "poktroll.proof.EventClaimCreated", events[0].Type)
 
 			event, err := cosmostypes.ParseTypedEvent(abci.Event(events[0]))
 			require.NoError(t, err)
