+++ x/proof/types/errors.go
@@ -31,4 +31,5 @@ var (
 	ErrProofParamInvalid              = sdkerrors.Register(ModuleName, 1120, "the provided param is invalid")
 	ErrProofClaimOutsideOfWindow      = sdkerrors.Register(ModuleName, 1121, "claim attempted outside of the session's claim window")
 	ErrProofProofOutsideOfWindow      = sdkerrors.Register(ModuleName, 1122, "proof attempted outside of the session's proof window")
+	ErrProofSupplierMismatch          = sdkerrors.Register(ModuleName, 1123, "supplier address does not match the claim or proof")
 )
